requirements.txt
aniso8601==8.1.1
astroid==2.4.2
boto3==1.15.0
botocore==1.18.18
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
imageio==2.9.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
networkx==2.5
numpy==1.19.5
opencv-python==4.2.0.32
Pillow==8.1.2
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
s3transfer==0.3.6
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.23
tifffile==2020.9.3
toml==0.10.2
typed-ast==1.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
